preliminaryProcessing: fix calculating I/Q squared for RMS mode
For RMS mode we need:

RMS = sqrt(A1^2 + ... + AN^2)

for I/Q, A = sqrt(I^2 + Q^2), so:

RMS = sqrt(I1^2 + Q1^2 + ... IN^2 + QN^2)

This commit fixes the inner part of the sqrt.
So, not the isqrt module could work correctly.

We still need to fix the correct samples to use
for this mode.